+++ jsk_pcl_ros/include/jsk_pcl_ros/delay_pointcloud.h
@@ -1,7 +1,8 @@
+// -*- mode: C++ -*-
 /*********************************************************************
  * Software License Agreement (BSD License)
  *
- *  Copyright (c) 2013, Yuto Inagaki and JSK Lab
+ *  Copyright (c) 2014, JSK Lab
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -32,46 +33,33 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  *********************************************************************/
 
+#ifndef JSK_PCL_ROS_DELAY_POINTCLOUD_H_
+#define JSK_PCL_ROS_DELAY_POINTCLOUD_H_
 
-#include "jsk_pcl_ros/tf_transform_cloud.h"
-#include <pluginlib/class_list_macros.h>
-
-#include <tf2_ros/buffer_client.h>
-#include <pcl/common/centroid.h>
+#include <pcl_ros/pcl_nodelet.h>
+#include <sensor_msgs/PointCloud.h>
 
 namespace jsk_pcl_ros
 {
-  void TfTransformCloud::transform(const sensor_msgs::PointCloud2ConstPtr &input)
-  {
-    sensor_msgs::PointCloud2 output;
-    try
-    {
-      if (pcl_ros::transformPointCloud(target_frame_id_, *input, output, tf_listener_)) {
-        pub_cloud_.publish(output);
-      }
-    }
-    catch (tf2::ConnectivityException &e)
-    {
-      NODELET_ERROR("Transform error: %s", e.what());
-    }
-    catch (tf2::InvalidArgumentException &e)
-    {
-      NODELET_ERROR("Transform error: %s", e.what());
-    }
-  }
 
-  void TfTransformCloud::onInit(void)
+  class DelayPointCloud: public pcl_ros::PCLNodelet
   {
-    PCLNodelet::onInit();
-    sub_cloud_ = pnh_->subscribe("input", 1, &TfTransformCloud::transform, this);
-    if (!pnh_->getParam("target_frame_id", target_frame_id_))
-    {
-      ROS_WARN("~target_frame_id is not specified, using %s", "/base_footprint");
-    }
+    
+  public:
+    
+  protected:
+    virtual void onInit();
+    virtual void delay(const sensor_msgs::PointCloud2::ConstPtr& msg);
 
-    pub_cloud_ = pnh_->advertise<sensor_msgs::PointCloud2>("output", 1);
-  }
+    double sleep_time_;
+    ros::Subscriber sub_;
+    ros::Publisher pub_;
+  private:
+
+  };
+
 }
 
-typedef jsk_pcl_ros::TfTransformCloud TfTransformCloud;
-PLUGINLIB_DECLARE_CLASS (jsk_pcl, TfTransformCloud, TfTransformCloud, nodelet::Nodelet);
+
+#endif
+