+++ jsk_pcl_ros/include/jsk_pcl_ros/depth_image_error.h
@@ -33,64 +33,41 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  *********************************************************************/
 
-#ifndef JSK_PCL_ROS_HANDLE_ESTIMATOR_H_
-#define JSK_PCL_ROS_HANDLE_ESTIMATOR_H_
 
+#ifndef JSK_PCL_ROS_DEPTH_IMAGE_ERROR_H_
+#define JSK_PCL_ROS_DEPTH_IMAGE_ERROR_H_
+
+#include <ros/ros.h>
 #include <pcl_ros/pcl_nodelet.h>
+#include <sensor_msgs/Image.h>
+#include <geometry_msgs/PointStamped.h>
+#include <jsk_pcl_ros/DepthErrorResult.h>
 
 #include <message_filters/subscriber.h>
 #include <message_filters/time_synchronizer.h>
 #include <message_filters/synchronizer.h>
+#include <message_filters/sync_policies/approximate_time.h>
 
-#include <tf/transform_listener.h>
-
-#include <jsk_pcl_ros/BoundingBox.h>
-#include "jsk_pcl_ros/Int32Stamped.h"
-#include <geometry_msgs/PoseArray.h>
-
-#include <boost/circular_buffer.hpp>
-#include <boost/tuple/tuple.hpp>
+#include <string>
 
 namespace jsk_pcl_ros
 {
-  class HandleEstimator: public pcl_ros::PCLNodelet
+  class DepthImageError: public pcl_ros::PCLNodelet
   {
   public:
-    typedef message_filters::sync_policies::ExactTime< sensor_msgs::PointCloud2,
-                                                       jsk_pcl_ros::BoundingBox > SyncPolicy;
-    enum HandleType
-    {
-      NO_HANDLE,
-      HANDLE_SMALL_ENOUGH_STAND_ON_PLANE,
-      HANDLE_SMALL_ENOUGH_LIE_ON_PLANE_Y_LONGEST,
-      HANDLE_SMALL_ENOUGH_LIE_ON_PLANE_X_LONGEST
-    };
-    
+    typedef message_filters::sync_policies::ApproximateTime<
+    sensor_msgs::Image,
+    geometry_msgs::PointStamped
+     > SyncPolicy;
+    ros::Publisher depth_error_publisher_;
+
   protected:
     virtual void onInit();
-    virtual void estimate(const sensor_msgs::PointCloud2::ConstPtr& cloud_msg,
-                          const jsk_pcl_ros::BoundingBox::ConstPtr& box_msg);
-    virtual void estimateHandle(const HandleType& handle_type,
-                                const sensor_msgs::PointCloud2::ConstPtr& cloud_msg,
-                                const jsk_pcl_ros::BoundingBox::ConstPtr& box_msg);
-    virtual void handleSmallEnoughLieOnPlane(
-      const sensor_msgs::PointCloud2::ConstPtr& cloud_msg,
-      const jsk_pcl_ros::BoundingBox::ConstPtr& box_msg,
-      bool y_longest);
-    virtual void handleSmallEnoughStandOnPlane(
-      const sensor_msgs::PointCloud2::ConstPtr& cloud_msg,
-      const jsk_pcl_ros::BoundingBox::ConstPtr& box_msg);
-
-    virtual void selectedIndexCallback( const jsk_pcl_ros::Int32StampedConstPtr &index);
-    ros::Publisher pub_, pub_best_, pub_preapproach_, pub_selected_, pub_selected_preapproach_;
-    ros::Subscriber sub_index_;
+    virtual void calcError(const sensor_msgs::Image::ConstPtr& depth_image,
+                           const geometry_msgs::PointStamped::ConstPtr& uv_point);
+    message_filters::Subscriber<sensor_msgs::Image> sub_image_;
+    message_filters::Subscriber<geometry_msgs::PointStamped> sub_point_;
     boost::shared_ptr<message_filters::Synchronizer<SyncPolicy> >sync_;
-    message_filters::Subscriber<sensor_msgs::PointCloud2> sub_input_;
-    message_filters::Subscriber<jsk_pcl_ros::BoundingBox> sub_box_;
-    boost::shared_ptr<tf::TransformListener> tf_listener_;
-    double gripper_size_;
-    double approach_offset_;
-    boost::circular_buffer<boost::tuple<geometry_msgs::PoseArray, geometry_msgs::PoseArray> > output_buf;
   private:
   };
 }