+++ jsk_pcl_ros/include/jsk_pcl_ros/grid_map.h
@@ -45,6 +45,7 @@
 #include <set>
 #include <Eigen/Geometry>
 
+#include "jsk_pcl_ros/geo_util.h"
 
 namespace jsk_pcl_ros
 {
@@ -80,6 +81,12 @@ namespace jsk_pcl_ros
     // toMsg does not fill header, be carefull
     virtual void originPose(Eigen::Affine3d& output);
     virtual void toMsg(SparseOccupancyGrid& grid);
+    virtual Plane toPlane();
+    virtual void vote();
+    virtual unsigned int getVoteNum();
+    virtual void setGeneration(unsigned int generation);
+    virtual unsigned int getGeneration();
+    virtual std::vector<float> getCoefficients();
   protected:    
     double resolution_;
     Eigen::Vector3f O_;
@@ -92,6 +99,8 @@ namespace jsk_pcl_ros
     
     std::vector<GridLine::Ptr> lines_;
     Columns data_;
+    unsigned int vote_;
+    unsigned int generation_;
   private:
   };