+++ jsk_pcl_ros/src/delay_pointcloud_nodelet.cpp
@@ -1,7 +1,8 @@
+// -*- mode: C++ -*-
 /*********************************************************************
  * Software License Agreement (BSD License)
  *
- *  Copyright (c) 2013, Yuto Inagaki and JSK Lab
+ *  Copyright (c) 2014, JSK Lab
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -32,46 +33,27 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  *********************************************************************/
 
-
-#include "jsk_pcl_ros/tf_transform_cloud.h"
+#include "jsk_pcl_ros/delay_pointcloud.h"
 #include <pluginlib/class_list_macros.h>
-
-#include <tf2_ros/buffer_client.h>
-#include <pcl/common/centroid.h>
 
 namespace jsk_pcl_ros
 {
-  void TfTransformCloud::transform(const sensor_msgs::PointCloud2ConstPtr &input)
-  {
-    sensor_msgs::PointCloud2 output;
-    try
-    {
-      if (pcl_ros::transformPointCloud(target_frame_id_, *input, output, tf_listener_)) {
-        pub_cloud_.publish(output);
-      }
-    }
-    catch (tf2::ConnectivityException &e)
-    {
-      NODELET_ERROR("Transform error: %s", e.what());
-    }
-    catch (tf2::InvalidArgumentException &e)
-    {
-      NODELET_ERROR("Transform error: %s", e.what());
-    }
-  }
-
-  void TfTransformCloud::onInit(void)
+  void DelayPointCloud::onInit()
   {
     PCLNodelet::onInit();
-    sub_cloud_ = pnh_->subscribe("input", 1, &TfTransformCloud::transform, this);
-    if (!pnh_->getParam("target_frame_id", target_frame_id_))
-    {
-      ROS_WARN("~target_frame_id is not specified, using %s", "/base_footprint");
-    }
-
-    pub_cloud_ = pnh_->advertise<sensor_msgs::PointCloud2>("output", 1);
+    pnh_->param("sleep_time", sleep_time_, 1.0);
+    sub_ = pnh_->subscribe("input", 1, &DelayPointCloud::delay, this);
+    pub_ = pnh_->advertise<sensor_msgs::PointCloud2>("output", 1);
   }
+
+  void DelayPointCloud::delay(const sensor_msgs::PointCloud2::ConstPtr& msg)
+  {
+    ros::Duration(sleep_time_).sleep();
+    pub_.publish(msg);
+  }
+
+
 }
 
-typedef jsk_pcl_ros::TfTransformCloud TfTransformCloud;
-PLUGINLIB_DECLARE_CLASS (jsk_pcl, TfTransformCloud, TfTransformCloud, nodelet::Nodelet);
+typedef jsk_pcl_ros::DelayPointCloud DelayPointCloud;
+PLUGINLIB_DECLARE_CLASS (jsk_pcl, DelayPointCloud, DelayPointCloud, nodelet::Nodelet);